Move src/utils/node to src/gatsby
