+++ src/gatsby/common.js
@@ -3,7 +3,7 @@ const process = require('process')
 const { createFilePath } = require('gatsby-source-filesystem')
 const remark = require('remark')
 const remarkHTML = require('remark-html')
-const { getItemBySource } = require('../shared/sidebar')
+const { getItemBySource } = require('../utils/shared/sidebar')
 
 const markdownToHtml = remark().use(remarkHTML).processSync
 const is404Regexp = /^\/404/
