+++ src/gatsby/doc.js
